Python/flask/flask_dio/pyproject.toml
[project]
name = "flask-dio"
version = "0.1.0"
description = "BLOG api"
authors = [
    {name = "spork1on",email = "diegobavutti@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "flask-jwt-extended (>=4.7.1,<5.0.0)",
]

[tool.poetry]
packages = [{include = "flask_dio", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
